# back/src/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Pagemaker API.
  version: 0.1.0
servers:
  - url: http://localhost:4200/api/v1.0
    description: Pagemaker API V1.0
paths:
...
  /sites:
    get:
      tags:
        - public
        - user
        - sites
      summary: get list of the users sites
      description: returns a list of the sites for q given user
      parameters:
        - name: userId
          in: path
          required: true
          description: logged in user id
          schema:
            type: string
      responses:
        '200':
          description: array containing the list of sites or empty array if none have been created
          content: 
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/site'
...
  /sites/{userId}/{siteId}:
    post:
      tags:
        - sites
      summary: Save details of a new site
      requestBody:
        description: details of the site
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/site'
      responses:
        '201':
          description: Site was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/site'
    put:
      tags:
        - sites
      summary: Save details of a new site
      requestBody:
        description: details of the site
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/site'
      responses:
        '201':
          description: Site was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/site'
...
components:
...
  schemas:
...
    site:
      type: object
      title: A site
      properties:
        siteId:
          title: unique id for the site
          type: string
        userId:
          title: id of the user
          type: string
        description:
          title: description of the site
          type: string
        image: 
          title: url pointing to an image linked to the site
          type: string
        name: 
          title: The name of the site
          type: string
...